refactor UI
