+++ src/app/components/policy-purchase-insured-info/policy-purchase-insured-info.component.html
@@ -111,9 +111,9 @@
               <nx-switcher formControlName="isUsPerson" labelSize="small"></nx-switcher>
               <nx-popover #usPersonTooltip>
                 @if (usPersonError) {
-                    <div style="max-width: 300px">
-                      This application is only open to non-U.S. persons. For more info, contact PDD.Support&#64;Allianz.com.my
-                    </div>
+                  <div style="max-width: 300px">
+                    This application is only open to non-U.S. persons. For more info, contact PDD.Support&#64;Allianz.com.my
+                  </div>
                 }
               </nx-popover>
             </div>
