+++ src/app/components/policy-purchase/policy-purchase.component.html
@@ -80,7 +80,9 @@
           }
 
           @case (3) {
-            Receipt
+            <div class="particulars-container">
+              <app-policy-purchase-summary></app-policy-purchase-summary>
+            </div>
           }
         }
       }
